Refine IGV DRS URI processing to GCS media API URLs
